requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-phonenumber-field==4.0.0
django-registration==3.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.5
openapi-codec==1.3.2
packaging==20.4
phonenumbers==8.12.5
Pillow==7.1.2
psycopg2==2.8.5
pyparsing==2.4.7
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
